Create a separate list folder
