+++ client/web/src/enterprise/campaigns/preview/list/ChangesetSpecNode.scss
