+++ client/web/src/enterprise/campaigns/preview/list/ChangesetSpecNode.story.tsx
@@ -4,7 +4,7 @@ import { ChangesetSpecNode } from './ChangesetSpecNode'
 import { visibleChangesetSpecStories } from './VisibleChangesetSpecNode.story'
 import { hiddenChangesetSpecStories } from './HiddenChangesetSpecNode.story'
 import { of } from 'rxjs'
-import { EnterpriseWebStory } from '../../components/EnterpriseWebStory'
+import { EnterpriseWebStory } from '../../../components/EnterpriseWebStory'
 
 const { add } = storiesOf('web/campaigns/apply/ChangesetSpecNode', module).addDecorator(story => (
     <div className="p-3 container web-content changeset-spec-list__grid">{story()}</div>
