+++ client/web/src/enterprise/campaigns/preview/list/ChangesetSpecNode.tsx
@@ -1,10 +1,10 @@
 import * as H from 'history'
 import React from 'react'
-import { ChangesetApplyPreviewFields } from '../../../graphql-operations'
+import { ChangesetApplyPreviewFields } from '../../../../graphql-operations'
 import { HiddenChangesetSpecNode } from './HiddenChangesetSpecNode'
 import { VisibleChangesetSpecNode } from './VisibleChangesetSpecNode'
-import { ThemeProps } from '../../../../../shared/src/theme'
-import { queryChangesetSpecFileDiffs } from './backend'
+import { ThemeProps } from '../../../../../../shared/src/theme'
+import { queryChangesetSpecFileDiffs } from '../backend'
 
 export interface ChangesetSpecNodeProps extends ThemeProps {
     node: ChangesetApplyPreviewFields
