+++ client/web/src/enterprise/campaigns/preview/list/EmptyChangesetSpecListElement.scss
