+++ client/web/src/enterprise/campaigns/preview/list/GitBranchChangesetDescriptionInfo.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useState } from 'react'
-import { GitBranchChangesetDescriptionFields } from '../../../graphql-operations'
-import { formatPersonName, PersonLink } from '../../../person/PersonLink'
-import { UserAvatar } from '../../../user/UserAvatar'
+import { GitBranchChangesetDescriptionFields } from '../../../../graphql-operations'
+import { formatPersonName, PersonLink } from '../../../../person/PersonLink'
+import { UserAvatar } from '../../../../user/UserAvatar'
 
 const dotDotDot = '\u22EF'
