+++ client/web/src/enterprise/campaigns/preview/list/VisibleChangesetSpecNode.scss
