+++ client/web/src/enterprise/campaigns/preview/list/VisibleChangesetSpecNode.tsx
@@ -1,16 +1,16 @@
 import * as H from 'history'
 import React, { useCallback, useState } from 'react'
-import { VisibleChangesetApplyPreviewFields, VisibleChangesetSpecFields } from '../../../graphql-operations'
-import { ThemeProps } from '../../../../../shared/src/theme'
-import { queryChangesetSpecFileDiffs as _queryChangesetSpecFileDiffs } from './backend'
-import { Link } from '../../../../../shared/src/components/Link'
-import { DiffStat } from '../../../components/diff/DiffStat'
+import { VisibleChangesetApplyPreviewFields, VisibleChangesetSpecFields } from '../../../../graphql-operations'
+import { ThemeProps } from '../../../../../../shared/src/theme'
+import { queryChangesetSpecFileDiffs as _queryChangesetSpecFileDiffs } from '../backend'
+import { Link } from '../../../../../../shared/src/components/Link'
+import { DiffStat } from '../../../../components/diff/DiffStat'
 import { ChangesetSpecAction } from './ChangesetSpecAction'
 import ChevronDownIcon from 'mdi-react/ChevronDownIcon'
 import ChevronRightIcon from 'mdi-react/ChevronRightIcon'
-import { FileDiffConnection } from '../../../components/diff/FileDiffConnection'
-import { FileDiffNode } from '../../../components/diff/FileDiffNode'
-import { FilteredConnectionQueryArguments } from '../../../components/FilteredConnection'
+import { FileDiffConnection } from '../../../../components/diff/FileDiffConnection'
+import { FileDiffNode } from '../../../../components/diff/FileDiffNode'
+import { FilteredConnectionQueryArguments } from '../../../../components/FilteredConnection'
 import { GitBranchChangesetDescriptionInfo } from './GitBranchChangesetDescriptionInfo'
 import { map } from 'rxjs/operators'
